+++ lib/primitive/lib/Cardano/Wallet/Primitive/Ledger/Read/Tx/Features/WitnessCount.hs
@@ -15,6 +15,10 @@ import Cardano.Ledger.Api
     , hashScript
     , scriptTxWitsL
     )
+import Cardano.Read.Ledger.Tx.Outputs
+    ( Outputs (..)
+    , getEraOutputs
+    )
 import Cardano.Wallet.Primitive.Ledger.Read.Tx.Features.Outputs
     ( fromBabbageTxOut
     , fromConwayTxOut
@@ -40,10 +44,6 @@ import Cardano.Wallet.Read
     , Mary
     , Shelley
     , Tx
-    )
-import Cardano.Wallet.Read.Tx.Outputs
-    ( Outputs (..)
-    , getEraOutputs
     )
 import Cardano.Wallet.Read.Tx.Witnesses
     ( Witnesses (..)
