+++ lib/read/cardano-wallet-read.cabal
@@ -45,6 +45,7 @@ library
   exposed-modules:
     Cardano.Read.Ledger
     Cardano.Read.Ledger.Tx.Inputs
+    Cardano.Read.Ledger.Tx.Outputs
     Cardano.Read.Ledger.Tx.TxId
     Cardano.Wallet.Read
     Cardano.Wallet.Read.Block
@@ -90,7 +91,6 @@ library
     Cardano.Wallet.Read.Tx.Integrity
     Cardano.Wallet.Read.Tx.Metadata
     Cardano.Wallet.Read.Tx.Mint
-    Cardano.Wallet.Read.Tx.Outputs
     Cardano.Wallet.Read.Tx.ReferenceInputs
     Cardano.Wallet.Read.Tx.ScriptValidity
     Cardano.Wallet.Read.Tx.TxId
